+++ core/API/DataTableManipulator/Flattener.php
@@ -63,15 +63,24 @@ class Flattener extends DataTableManipulator
     {
         $newDataTable = $dataTable->getEmptyClone($keepFilters = true);
 
-        foreach ($dataTable->getRows() as $rowId => $row) {
-            $this->flattenRow($row, $rowId, $newDataTable);
-        }
+        // this recursive filter will be applied to subtables
+        $dataTable->filter('ReplaceSummaryRowLabel');
+        $dataTable->filter('ReplaceColumnNames');
 
-        $queuedFiltersEnabled = Common::getRequestVar('disable_queued_filters', 0, 'int', $this->request) == 0;
-        if ($queuedFiltersEnabled) {
-            $newDataTable->applyQueuedFilters();
-        }
+        $this->flattenDataTableInto($dataTable, $newDataTable);
+
         return $newDataTable;
+    }
+
+    /**
+     * @param $dataTable DataTable
+     * @param $newDataTable
+     */
+    protected function flattenDataTableInto($dataTable, $newDataTable, $prefix = '', $logo = false)
+    {
+        foreach ($dataTable->getRows() as $rowId => $row) {
+            $this->flattenRow($row, $rowId, $newDataTable, $prefix, $logo);
+        }
     }
 
     /**
@@ -128,9 +137,7 @@ class Flattener extends DataTableManipulator
                 $dataTable->addRow($row);
             }
             $prefix = $label . $this->recursiveLabelSeparator;
-            foreach ($subTable->getRows() as $rowId => $row) {
-                $this->flattenRow($row, $rowId, $dataTable, $prefix, $logo);
-            }
+            $this->flattenDataTableInto($subTable, $dataTable, $prefix, $logo);
         }
     }
 
@@ -146,4 +153,5 @@ class Flattener extends DataTableManipulator
 
         return $request;
     }
+
 }