+++ core/Db/Schema/Mysql.php
@@ -35,7 +35,7 @@ class Mysql implements SchemaInterface
         $tables = array(
             'user'    => "CREATE TABLE {$prefixTables}user (
                           login VARCHAR(100) NOT NULL,
-                          password CHAR(32) NOT NULL,
+                          password VARCHAR(255) NOT NULL,
                           alias VARCHAR(45) NOT NULL,
                           email VARCHAR(100) NOT NULL,
                           token_auth CHAR(32) NOT NULL,