+++ plugins/ScheduledReports/API.php
@@ -15,6 +15,9 @@ use Piwik\Config;
 use Piwik\Container\StaticContainer;
 use Piwik\Date;
 use Piwik\Db;
+use Piwik\Development;
+use Piwik\Filesystem;
+use Piwik\Http;
 use Piwik\Log;
 use Piwik\NoAccessException;
 use Piwik\Piwik;
@@ -395,7 +398,7 @@ class API extends \Piwik\Plugin\API
             } catch (\Exception $ex) {
                 // NOTE: can't use warning or error because the log message will appear in the UI as a notification
                 $this->logger->info("Error getting '?{report}' when generating scheduled report: {exception}", array(
-                    'report' => http_build_query($params),
+                    'report' => Http::buildQuery($params),
                     'exception' => $ex->getMessage(),
                 ));
 
@@ -609,10 +612,9 @@ class API extends \Piwik\Plugin\API
         $now = Date::now()->getDatetime();
         $this->getModel()->updateReport($report['idreport'], array('ts_last_sent' => $now));
 
-        // If running from piwik.php with debug, do not delete the PDF after sending the email
-        $tracker = new Tracker();
-        if (!$tracker->isDebugModeEnabled()) {
+        if (!Development::isEnabled()) {
             @chmod($outputFilename, 0600);
+            Filesystem::deleteFileIfExists($outputFilename);
         }
     }