+++ plugins/WebsiteMeasurable/MeasurableSettings.php
@@ -8,6 +8,7 @@
 
 namespace Piwik\Plugins\WebsiteMeasurable;
 use Piwik\IP;
+use Piwik\Measurable\Type\TypeManager;
 use Piwik\Network\IPUtils;
 use Piwik\Piwik;
 use Piwik\Plugin;
@@ -16,7 +17,6 @@ use Piwik\Settings\Setting;
 use Piwik\Settings\FieldConfig;
 use Piwik\Plugins\SitesManager;
 use Exception;
-use Piwik\Url;
 
 /**
  * Defines Settings for ExampleSettingsPlugin.
@@ -74,16 +74,38 @@ class MeasurableSettings extends \Piwik\Settings\Measurable\MeasurableSettings
      */
     private $pluginManager;
 
-    public function __construct(SitesManager\API $api, Plugin\Manager $pluginManager, $idSite, $idMeasurableType)
+    /**
+     * @var TypeManager
+     */
+    private $typeManager;
+
+    public function __construct(SitesManager\API $api, Plugin\Manager $pluginManager, TypeManager $typeManager, $idSite, $idMeasurableType)
     {
         $this->sitesManagerApi = $api;
         $this->pluginManager = $pluginManager;
+        $this->typeManager = $typeManager;
 
         parent::__construct($idSite, $idMeasurableType);
     }
 
+    protected function shouldShowSettingsForType($type)
+    {
+        $isWebsite = $type === Type::ID;
+
+        if ($isWebsite) {
+            return true;
+        }
+
+        // if no such type exists, we default to website properties
+        return !$this->typeManager->isExistingType($type);
+    }
+
     protected function init()
     {
+        if (!$this->shouldShowSettingsForType($this->idMeasurableType)) {
+            return;
+        }
+
         $this->urls = new Urls($this->idSite);
         $this->addSetting($this->urls);
 
@@ -183,7 +205,7 @@ class MeasurableSettings extends \Piwik\Settings\Measurable\MeasurableSettings
         $self = $this;
         return $this->makeProperty('excluded_parameters', $default = array(), FieldConfig::TYPE_ARRAY, function (FieldConfig $field) use ($self) {
             $field->title = Piwik::translate('SitesManager_ExcludedParameters');
-            $field->inlineHelp = Piwik::translate('SitesManager_ListOfQueryParametersToExclude')
+            $field->inlineHelp = Piwik::translate('SitesManager_ListOfQueryParametersToExclude', "/^sess.*|.*[dD]ate$/")
                 . '<br /><br />'
                 . Piwik::translate('SitesManager_PiwikWillAutomaticallyExcludeCommonSessionParameters', array('phpsessid, sessionid, ...'));
             $field->uiControl = FieldConfig::UI_CONTROL_TEXTAREA;