+++ tests/PHPUnit/Integration/HttpTest.php
@@ -55,6 +55,13 @@ class HttpTest extends \PHPUnit_Framework_TestCase
         $this->assertGreaterThan(0, filesize($destinationPath));
     }
 
+    public function testBuildQuery()
+    {
+        $this->assertEquals('', Http::buildQuery(array()));
+        $this->assertEquals('test=foo', Http::buildQuery(array('test' => 'foo')));
+        $this->assertEquals('test=foo&bar=baz', Http::buildQuery(array('test' => 'foo', 'bar' => 'baz')));
+    }
+
     /**
      * @dataProvider getMethodsToTest
      */
@@ -230,12 +237,17 @@ class HttpTest extends \PHPUnit_Framework_TestCase
     }
 
     /**
+     * erroe message can be:
+     *      curl_exec: server certificate verification failed. CAfile: /home/travis/build/piwik/piwik/core/DataFiles/cacert.pem CRLfile: none. Hostname requested was: self-signed.badssl.com
+     * or
+     *      curl_exec: SSL certificate problem: self signed certificate. Hostname requested was: self-signed.badssl.com
      * @expectedException \Exception
-     * @expectedExceptionMessage curl_exec: SSL
+     * @expectedExceptionMessageRegExp /curl_exec: .*certificate.* /
      */
     public function testCurlHttpsFailsWithInvalidCertificate()
     {
-        Http::sendHttpRequestBy('curl', 'https://www.virtual-drums.com', 10);
+        // use a domain from https://badssl.com/
+        Http::sendHttpRequestBy('curl', 'https://self-signed.badssl.com/', 10);
     }
 
     /**
@@ -244,7 +256,8 @@ class HttpTest extends \PHPUnit_Framework_TestCase
      */
     public function testFopenHttpsFailsWithInvalidCertificate()
     {
-        Http::sendHttpRequestBy('fopen', 'https://www.virtual-drums.com', 10);
+        // use a domain from https://badssl.com/
+        Http::sendHttpRequestBy('fopen', 'https://self-signed.badssl.com/', 10);
     }
 
     /**